requirements.txt
Django==1.11.5
dj-database-url==0.4.2
djangorestframework==3.7.0
gunicorn==19.7.1
psycopg2==2.7.3.1
pytz==2017.2
whitenoise==3.3.0
